# src/docs/openapi.yaml
openapi: 3.0.3  
info:
  title: Price API
  description: API for Price Aggregator Project
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        amountUsdt:
          type: number
        amountCurrency:
          type: number
        currency:
          type: string
        transactionId:
          type: string
        status:
          type: string
          enum: [pending, confirmed, failed]
        createdAt:
          type: string
          format: date-time
    ApiKey:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        key:
          type: string
        createdAt:
          type: string
          format: date-time
        lastUsedAt:
          type: string
          format: date-time
security:
  - BearerAuth: []
...
paths:
...
  /payments/verify:
    post:
      tags: [Payments]
      summary: Confirm user payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [transactionId, amountCurrency, currency, plan,walletAddress,amountUsdt]
              properties:
                transactionId:
                  type: string
                  example: 0x123abc...
                amountCurrency:
                  type: number
                  example: 100 
                amountUsdt:
                  type: number
                  example: 100
                currency:
                  type: string
                  example: USDT
                plan:
                  type: string
                  enum: [basic, growth, pro]
                walletAddress:
                  type: string

      responses:
        200:
          description: Payment confirmed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Payment confirmed and plan updated
                  payment:
                    $ref: '#/components/schemas/Payment'
  /payments/upgrade:
    post:
      tags: [Payments]
      summary: upgrade user plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [transactionId, currency, plan,walletAddress,amountUsdt]
              properties:
                transactionId:
                  type: string
                  example: 0x123abc...
                amountCurrency:
                  type: number
                  example: 100 
                amountUsdt:
                  type: number
                  example: 100
                currency:
                  type: string
                  example: USDT
                plan:
                  type: string
                  enum: [basic, growth, pro]
                walletAddress:
                  type: string

      responses:
        200:
          description: Payment confirmed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Payment confirmed and plan updated
                  payment:
                    $ref: '#/components/schemas/Payment'
...
  /api-keys:
    post:
      tags: [API Keys]
      summary: Generate new API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  example: MyKey1
      responses:
        201:
          description: API key created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
    get:
      tags: [API Keys]
      summary: Get all API keys for current user
      responses:
        200:
          description: API keys retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiKey'
  /api-keys/{id}:
    delete:
      tags: [API Keys]
      summary: Delete an API key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: API key deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: API key deleted successfully

  # ----------------- User Profile APIs -----------------
...